Use expanding bindparam in provider_ids_from_rp_ids in_
This was explored as part of discussion with zzzeek on various
ways to make a statement with a large in_ be more performant.

Batching the queries by chunking rp_ids did not help, but an
expanding bindparam did: it makes the surrounding method about
50% faster [1].

bindparam expanding was added in sqlalchemy 1.2.0 so we bump
the requirements here to the latest 1.2.x bugfix release,
1.2.19.

[1] With the caveat that profiling and benchmarking are very
different things, and in this case it is profiling results
which are being analysed.

Change-Id: Ic4e0cdd87f8f2d76b921059ac4bf16a838913abf