Use __slots__ in commonly used objects
When making a nested provider GET /allocation_candidates request
that returns around 6500 providers in the providers summary section,
we make ~22000 ResourceProvider and ProviderSummary objects, 7500
AllocationRequest objects and 30000 ProviderSummaryResource objects.

In aggregate, creating and accessing these objects contributes a
measurable cpu time cost. Using __slots__ [1] can speed up attribute
access (I measured it, it does), so lets use it for the common
objects. There's no evidence (yet) of other objects being an issue
so we won't do this wholesale.

A unit test is updated to change a mock from a public ResourceProvider
method (set_aggregates) to a private function (_set_aggregates)
because one of the limitations of __slots__ is it makes the
slotted objects difficult to mock. In this case the test is still
testing the same thing.

Subsequent patches will explore why we need 22000 ProviderSummary
objects when we're only producing 6500 resulting summaries to see
if there is a cleanup that can be done there.

[1] https://docs.python.org/3/reference/datamodel.html#slots

Change-Id: I4acb16d13976d64d47f798ebbec626a73b79b4f5