Add RequestWideSearchContext.summaries_by_id
When making a GET /allocation_candidates request, ProviderSummaries
are created for each request group. This can lead to duplicate calls
to creating the ProviderSummary objects and the ResourceProvider
object each contains.

The usage info is fixed for the duration of the request so there is
no difference between ProviderSummaries with the same id

Therefore a summaries_by_id dict is added to the
RequestWideSearchContext to operate as a single-request-wide cache of
ProviderSummary objects so that we don't need to create duplicates.

In testing using a large number of providers modeled after an
enhanced nested perfload topology [1], adding the cache more than
halfed the number of ProviderSummary and ResourceProvider objects
created. Since creating objects in Python is expensive, this is a
a win.

[1] I67d8091b464cd7b875b37766f52818a5a2faa780

Change-Id: I43ae1118421366336b4e96738c2981e07caebec8