Make placement base API return version without auth
Allow both GET '/placement' and '/placement/' to return
version info without doing authentication.

In many deployment placement is found on a prefix of '/placement'
different servers will or will not append the '/'. To be most
resilient, accept either as "root".

A unit test is added which confirms that no auth is required
for root but auth is still required elsewhere.

Co-Authored-By: Chris Dent <cdent@anticdent.org>
Related-Bug: #1838633
Change-Id: Iaa93ec943f4f4fbb6c1af3a809859813872e62d7