Track usage info on RequestWideSearchContext
We were building usage information for each RequestGroup. This
could often involve looking up usage information for duplicated
providers.

Now, track the seen root ids on the RequestWideSearchContext in
a set, and extend a list of usages (on that object) only by those
root ids that we haven't yet seen.

In the process move _get_usages_by_provider_tree to
RequestWideSearchContext and rename it to
extend_usages_by_provider_tree

Manual testing indicates that after the first call to extend... we
can frequently skip it entirely.

Change-Id: I5d05d15d33beb1622a06bde0e32c3f52ba01b448