Move provider_ids_from_rp_ids to allocation_candidate and fix
Right now the only consumer of provider_ids_from_rp_ids is
_build_summaries, which is private to allocation_candidates.

Since provider_ids_from_rp_ids has evolved to have a fairly
specific behavior, with some tricky considerations for how
it is used, this patch moves it to allocation_candidate.py
and makes it private to make it a bit more clear that it is
special.

If/when it gains more uses we can reconsider.

This change also returns the expanding bindparam functionality
added in [1] and removed in [2] as part of a rebasing mistake.
See above about "tricky".

[1] Ic4e0cdd87f8f2d76b921059ac4bf16a838913abf
[2] If937053a8af3f0eecdd90aa807be0abc316d8c4f

Change-Id: Ief3e86ca0d722035809115818d38fb14a89a7a39