Add __copy__ method to AllocationRequest{,Resource}
It turns out that when Python runs copy.copy on an object it does
a lot of unnecessary work that can be avoided by implementing a
__copy__ method that essentially does a clone(), creating a new
object with the same values. This is considered the idiomatic form
of clone.

That unnecessary work shows up quite significantly in the
_consolidate_allocation_requests method in allocation_candidate.py where
AllocationRequestResource objects get copied, a lot. The use of __copy__
more than halves the time consumed by _consolidate_allocation_requests.

There is another copy.copy, in _alloc_candidates_single_provider in the
same file, this time copying AllocationRequest objects. This is only
done when a shared provider is involved, and we haven't been doing scale
testing for those (yet) but since we're in that domain I've gone ahead
and pre-emptively added a __copy__ to that object too, to prepare for
the eventual work of many shared disks.

Change-Id: I3e8167e5d09aeb2ae68282bc0378bee6d956a286