Copy AllocationRequestResource only when necessary
Profiling revealed that copying AllocationRequestResource in
_consolidate_allocation_requests was expensive.

As an optimization, this patch adds logic to only do that copy when it's
necessary.

And it's only necessary for an ARR whose resource class was requested
from multiple groups, when group_policy=none. In that case, we're
constructing multiple results with the same ARR. For example, consider a
host with two PFs with VF inventory. A request like:

 ?resources1=VF:1
 &resources2=VF:1
 &group_policy=none

...will yield results like:

(1) PF1{VF:1},PF2{VF:1}
(2) PF1{VF:2}
(3) PF2{VF:2}

The AllocationRequestResource representing PF1{VF:1} gets used in both
result (1) and result (2). When producing result (2), we add the
AllocationRequestResource.amountZ together. If we were reusing the same
ARR instance from result (1), result (1) would end up looking like:

(1) PF1{VF:2},PF2{VF:1}

...which is wrong. So we have to do the copy.

With group_policy=isolate, or if we've only requested the resource class
from one group, we're only using that ARR once in the results, so we can
just use it without copying.

Change-Id: Idb98611f549b628d273c43f07d137fcf9c73314c