Add rw_ctx.parent_uuid_by_rp_uuid, for clarity
In _merge_candidates a dict of rp uuid to parent uuid mappings
is used to determine if providers are in the same subtree.

In this change the creation of that dict is moved from _merge_canidates
to _build_provider_summaries because we have established that all
relevant providers will be visited by the sum of calls to that method
and we do not want to have to rely on visiting all providers in
_merge_candidates, to allow for the possibility of short circuiting
some loops.

The signature on _satisfies_same_subtree is updated to accept an
rw_ctx, rather than two separate members of the object.

Change-Id: Iaba4a098847dc4a6fc1bdaee44971dd4f0e051ae