Add a rw_ctx.psum_res_by_rp_rc, for clarity
In _merge_candidates we need a final _exceeds_capacity check to
make sure a multi-part granular request has not exceeded capacity
limits. This uses a dict that maps resource provider uuid and resource
class name to a ProviderSummaryResource.

Rather than creating this dict in _merge_candidates, we now have it
as a member on the RequestWideSearchContext and add to it when
creating the ProviderSummaryResource.

This allows us to remove some looping in _merge_candidates and is
also a bit more tidy. This is only possible with the advent of
the RequestWideSearchContext, which is newer than _merge_candidates.

_exceeds_capacity is moved to the RequestWideSearchContext as
exceeds_capacity as it now makes sense as a method.

Because _rp_rc_key ended up used from two different modules and for
two different purposes, the method is instead removed and callers
create their own keys directly.

Change-Id: Id74c01215956fc998f2dadcd9d84801ac58c2d3d