Use rp.id instead of uuid in _rp_rc_key
This is the tiniest of memory and hashing optimizations but is
also for subsequent changes which will change when and where a
key is create and used in the management of psum_res_by_rp_rc.

All the current callers have an rp with id set when they use this.

Change-Id: I9d8f4ed1c32deafe90cd4c4ddf7c36f11a56b6b5