Fix allocation bug in NUMANetworkFixture
While doing something else, I noticed a bug in the NUMANetworkFixture:
we were trying to consume two VCPU from numa0 and three from cn2; but
our second call to tb.set_allocation() was using the same consumer as
the first, so numa0 was losing its allocation.

This commit uses a new consumer for the cn2 allocation, and tweaks an
existing gabbi test to prove the fix.

Change-Id: I893b27b7a07d20ef245c6056cde0e2c5bc5f4daa