gabbi test for same_subtree with an ancestry hole
Based on the analysis at [1], we need to continue to use fully-populated
ancestry data when calculating same_subtree-ness. This commit introduces
a gabbi test that covers a particular case that can be broken if we
don't do that: where a provider involved in the request has a broken
ancestry path to another provider involved in the request.

[1] https://review.opendev.org/#/c/675606/2/placement/objects/allocation_candidate.py@908

Change-Id: I5062e725d6f29cd997b974a0dad7dc4d0ed2900f