Fix section structure for pdf docs
In the document source, we had a hidden toctree to avoid an error
when creating an html document. However, the contents in the hidden
toctree appears explicitly as actual subsections. This resulted in
having wrong subsection structures when creating a pdf document.

This patch fixes it by adding explicit sections for the contents
in the (previously) hidden toctree.

Change-Id: I8420835c19856953c65513914362bf402ff0f08b
Story: 2006110
Task: 35398