Clean up document's index
In the top index page of the document, some sections had actual
contents and some had only links to the contents. However, this
makes it difficult for readers to get what is in the document and
what isn't at a glance.

This patch cleans it up to simplify the top index. This patch also
renames some of the directories to follow the doc layout rules[1].

[1] https://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Note that the following ToDo is left for further cleanup:

 * Publish the api guide and move the microversion history to
   the `/api-guide` directory

Change-Id: Ideb1d5e7c74d1d564d07bac1a1e7c86c55ede873