Add train-prelude release note
This is essentially a duplicate of the cycle highlights provided
with the release, plus a warning about needing to upgrade to extracted
placement in stein before the train upgrade.

Few links are provided because the rest of the release notes are very
short and will be presented adjacent to the prelude, and they have
links.

Change-Id: I872271688fbee69ff35bf44ef82fad9ab34f5229