Drop support for python 2
This is modelled on the similar change in nova [1] but is
simpler because Placement was already using python3 for
"integration" style tests.

The 'train' python3 template is replaced with corresponding 'ussuri'.

The python template is removed (it is 2.7) and the pep8 job that was
included in that template is now listed in check and gate jobs.

.zuul.yaml is updated to not do a generic 'funtional' test,
instead versioned tests only.

setup.cfg removes indications of support for python 2.

A reno is added.

tox.ini is changed to make the functional environment a
base environment for specific functional jobs.

[1] Ie1a0cbd82a617dbcc15729647218ac3e9cd0e5a9

Change-Id: Ifce3b0c0ff11553f32b2d02df07fde81efc34100