Bump oslo.log version to 4.3.0
This ensures we have a version available that contains the deprecated
markers for the Wallaby release.

This commit also updates the minimum version of oslo.context to 2.21.0.
The oslo.log library depends 2.20.0 and we'll need version 2.21.0 in a
subsequent patch.

I bumped additional packages to ensure the minimum versions for
lower-constraints work with the newer oslo dependencies.

Change-Id: I6d1e39886d177d8729bc2db9e0018b5c772f8b0e