Implement secure RBAC for resource providers
This commit updates the policies for the resource providers in placement to
support read-only roles.

This is part of a broader community effort to support read-only roles
and implement secure, consistent default policies.

This commit also introduces some testing infrastructure and plumbing
that is useful for implementing protection tests with gabbi. Including
testing for deprecated policies.

Change-Id: Icc904bf325eaa60377171a22a86932135c384e6a