placement-status: check only consumers in allocation table
Initially, _check_incomplete_consumers was intended to find the case
when consumers are mentioned in allocation but do not exist in the
consumer table. Due to a bug it's possible to have rows in the
consumer table those are not in the allocation table. The second case
it seems was not expected and it can be tricky to automate a cleanup
with zero touch upgrade.

This patch modifies the check such way it counts only consumers
presented in the allocation table.

Conflicts:
    placement/cmd/status.py

NOTE(melwitt): The conflict is because change
Ie0bee79c1ee185c0f68fa170c668395d9058c56d (Fix SQL query counting the
number of individual consumers having allocations by only selecting the
aggregated consumer_id column.) is not in Wallaby.

Related-bug: #1780799
Change-Id: I569043dd9f1c964faf7d4fcbb901a93f9569bf3d
(cherry picked from commit 3253e2efc05836b0b5cefafbf2b2656e9e1a50c7)