Pass context objects to oslo.policy directly
This oslo.policy Enforcer() object understands oslo.context
RequestContext objects and will translate the appropriate policy values
for you if it's passed into enforcement.

This lowers the bar for people interacting with oslo.policy because they
don't have to translate important authorization inforamtion from
RequestContext into a dictionary, specifically for policy enforcement.
Instead, they can just pass in the context object and oslo.policy will
know what to do.

This commit updates the policy module to do just that and moves the
credential tranlations to where it's needed, for logging.

Change-Id: I942da4b1880cf2312af2f8e15aef9f33be991f87