policy: Remove the deprecated 'placement' rule
The 'placement' rule pre-dates the granular policies introduced in the
Rocky release, back when placement was still part of nova. It's no
longer used anywhere and can/should be removed.

This commit also updates a policy test to use a registered policy,
instead of the `placement` rule we're removing here.  Otherwise, the
`authorize()` implementation will fail because the rule isn't officially
registered by placement.

Change-Id: I4b34c86a1991f954fc562b8e252fe9a178a08ad7
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>