policy: Don't persist default rule changes in tests
'oslo.policy' expects a list of default rules be provided when
initialising the enforcer. Unfortunately, it demonstrates a nasty habit
of modifying these defaults once it has them [1]. This can result in
configuration from one test bleeding through to the next.

Resolve this by initialising the enforcer with a copy of the default
rules, rather than the originals, allowing 'oslo.policy' to do whatever
it likes to the rules. Ultimately this should probably be fixed in
'oslo.policy', which should be either copying the defaults or creating
new rules based on these defaults, but that needs some thought and can
be done later.

[1] https://github.com/openstack/oslo.policy/blob/3.6.0/oslo_policy/policy.py#L762-L764

Change-Id: Ib864606d4cf4a8ecba95386bcc2b8a4bb68f2d17
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Story: 2008556
Task: 41687