Add weekly jobs
This change intoduces a set of jobs
that will run weekly in the periodic-weekly pipeline.
The intention of these jobs is to check the health
of the placement gate in the absence of frequent patch
submissions to ensure that new lib releases and or changes
in upper constraints or other placment related projects do
not result in broken jobs. The result of these jobs will
be review in our weekly meetings.

Change-Id: I456e44e69e503f2444ce9153aacee5783ea2b2eb