db: Replace implicit conversion of SELECT into FROM
Resolve the following warnings issued in recent versions of SQLAlchemy:

  SADeprecationWarning: Implicit coercion of SELECT and textual SELECT
  constructs into FROM clauses is deprecated; please call .subquery() on
  any Core select or ORM Query object in order to produce a subquery
  object.

The resolution is simple: instead of using 'FromClause.alias', use
'Query.subquery'. A warning filter is included to prevent us
reintroducing this issue in the future. It can be dropped once we start
using SQLAlchemy 2.0 where this behavior will be removed.

This change requires bumping SQLAlchemy to 1.4.0, which is the first
version to introduce the selectable '.subquery()' method. We must also
bump oslo.db to 8.6.0, which is the first version to support SQLAlchemy
1.4.x. The alternative would be to introduce a switch based on
SQLAlchemy versions, but that's tech debt that'll have to be cleaned up.

Change-Id: I6300be962dc3ce391f70d7c7e8af7cb4d6ce9baf
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>