Narrow scope of set allocations database transaction
When consumer types were added [1], the set allocations code path was
modified to perform util.ensure_consumer() and
AllocationList.replace_all() in a single database transaction in order
to provide automatic rollbacks of consumer type changes by racing
requests.

This proved to be problematic because util.ensure_consumer() contains
logic to "get or create" database records for project, user, consumer
type, and consumer. By definition, the "get or create" pattern cannot
reside in the same database transaction else a DuplicateEntry database
error followed by an attempt to "get" the existing record will fail as
the single database transaction is being rolled back and no new
statements can be executed.

Because of this, we cannot wrap util.ensure_consumer() under the same
database transaction as Allocation.replace_all() and we need to reduce
the scope of the database transaction to cover only the updates to a
consumer (if the requested project/user/consumer type differ from what
was found in the database) and the write of the allocation records.

This changes things around to stop performing "difference" updates
inside of the util.ensure_consumer() method and instead return what was
requested as an additional return value. Then, we move the database
transaction to cover only util.update_consumers() and
AllocationList.replace_all() so that if replace_all() fails due to a
generation conflict, the consumer updates will also be rolled back.

This approach was chosen in an effort to introduce the least amount of
change to the existing understood code.

Story: 2009159
Task: 43143

[1] I24c2315093e07dbf25c4fb53152e6a4de7477a51

Change-Id: I887606047ac1da3f2c59c17988ddcc17bd01ef43