Fix adding 'unknown' to the ConsumerTypeCache
The attribute cache added for consumer types in
I32499b2616e253a06f8a003fd3eef91006b58f42. It added the
`unknown` value to the cache as that is not persisted. But it
mixed up which internal dict represents which direction of
the cache and only populated one direction of the cache.
This is causes lookup errors for allocation with
consumer_type_id = None.

The cache is fixed now.

Story: 2009167
Task: 43183
Change-Id: I232acaad8b350b3cf924dd7a59d97eb8ad5d0fcc