Commonize code for updating and creating allocations
This is a followup patch for change
I887606047ac1da3f2c59c17988ddcc17bd01ef43 to refactor repeated code
into a common utility function. Note that the logic for deleting
consumers is kept separate because the consumer deletes need to be done
in a separate database transaction from the consumer updates and
allocation creates.

Change-Id: I9d02836c976140689c103e5b559e0221651907f8