Bump min decorator to 4.0.0
To be compatible with recent setuptools 58.0.0 which removed the support
for use_2to3 we need to bump the minimum version we support for
decorator from 3.4.0 to 4.0.0.

This is needed to unblock the gate.

From stable point of view such bump might not be allowed. Still stable/xena is not released yet so we can put this patch into RC2 and release it as part of the main release.

Change-Id: I456f8136f882a2c6621549b2b0f80e213e5430df
(cherry picked from commit 967abd1d57cfacb4ea276e0f909f379a59e2fce0)