Clarify trait filtering in the API doc
Based on the spec[1], the user doc[2] and the implementation[3][4]
the original intention of trait filtering during allocation
candidate query is that the required traits in the unnamed group
needs to be provided by the resource providers contributing resources
to the candidate. So traits on RPs which are in the tree of the
candidate but not providing resource to the candidate should be ignored.

The current API doc is not clear on this and can be misunderstood. So
this patch clarifies that doc.

[1] https://docs.openstack.org/placement/latest/specs/train/implemented/2005575-nested-magic-1.html
[2] https://docs.openstack.org/placement/latest/user/provider-tree.html#filtering-by-traits
[3] https://github.com/openstack/placement/blob/d48d22ff24535eb5ae947d60fd87a9afb2e7b45d/placement/objects/research_context.py#L745
[4] https://github.com/openstack/placement/blob/d48d22ff24535eb5ae947d60fd87a9afb2e7b45d/placement/tests/functional/db/test_allocation_candidates.py#L817-L827

Story: #2009795
Task: #44334
Change-Id: Idc367df609dc8d4874fc329f2317c582b9e06a4d