DB layer should only depend on trait id not names
Update get_provider_ids_having_any_trait to only depend on the internal
trait IDs in signature. In implementation it only depends on that anyhow
and all call sites already has access to the trait ids.

Later this change will help to use this call from call sites which only
have access to the trait ids but not the names any more.

Story: 2005345
Story: 2005346
Change-Id: Id1b3d6fb2c1a50909649583dbd8ca6e841468366