Fix perfload jobs after consumer_types
Since consumer_types was added to the API in
I24c2315093e07dbf25c4fb53152e6a4de7477a51 the two perfload jobs are
getting errors from placement as they are using the latest microversion
but does not specify the consumer_type when creating allocations.

The server could not comply with the request since it is either malformed
or otherwise incorrect.\n\n JSON does not validate: 'consumer_type' is a
required property

This patch changes the allocation request to specify TEST as consumer
type.

Change-Id: I31500e3e6df5717d6bdb6ed7ed43325653d49be5