db: Use Row, not LegacyRow
We added this check in change I8b845a29fa21f4ca8340d91bf5c6394094c6ab97
("db: Remove use of non-integer/slice indices") because we were using
the same code to handle two different types of object in the same
method. However, while the functions being called to generate this
function return a LegacyRow in recent versions of SQLAlchemy, it seems
older versions return a Row. The former is a subclass of the latter so
we can use the latter for both. This is also a better idea long-term,
since SQLAlchemy 2.0 removes LegacyRow.

Change-Id: Ia30c03b3878d80550cefd39271e315065f1a199d
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>