Make perfload jobs fail if write allocation fails
This uses curl -f when writing an allocation in order to detect when
the server has responded with a HTTP error code and then fails the job
if so. The idea behind this is to catch when PUT
/allocations/{consumer_uuid} required parameters change and the
perfload jobs need to be updated.

The curl -S option is also added to show the error if curl fails.

Change-Id: Ic06e64b1031ff37d7ada55449ae71cd39b1298a2