Add WA about resource_providers.can_host removal
We made a mistake and not added resource_providers.can_host to the new
DB schema of Placement when Placement was split out of Nova. This patch
documents that as can_host is not in the DB model it is safe to drop it
from the DB to remove the schema inconsistency.

Related-Bug: #1978549
Change-Id: Icfe83313e86969b0b9d34c19b6b82a47d685d4d7