Make us compatible with oslo.db 12.1.0
With oslo.db 12.1.0 the following sqlalchemy warning become an error:

  sqlalchemy.exc.RemovedIn20Warning: Retrieving row members using strings
  or other non-integers is deprecated; use row._mapping for a dictionary
  interface to the row (Background on SQLAlchemy 2.0 at:
  https://sqlalche.me/e/b8d9)

We tried to fix this before but missed a test_case that still used dict
access to get the fields of a Row instead of attribute access. We fixed
that test here. Also while fixed it I noticed that the generic
_AttributeCache object states that it stores dicts but actually it
sometimes stores dict but sometimes it stores Row objects. So the doc is
updated and the dict path converted to store namedtuple objects in the
cache instead. Note that Row is also acts like a namedtuple except that
._mapping does not exists in namedtuple but exists in Row. The trait
object assumed it gets a Row object with ._mapping from the cache so
that is adjusted to only assume a namedtuple and use _asdict() to covert
it to dict which is available both in Row and namedtuple.

Change-Id: I23ac1d85290a2dec307f8e76aafb02096259b605